INTEL_DII/NOT_UPSTREAM: drm/i915/vf: Add selftest subtest for GGTT region outside VF using PF
VF cannot read the contents of PTEs located outside its region, so
we are not sure if it is actually unable to modify them. The only
solution is to this via PF.
Let's add tests that verify via PF, that VF cannot modify an area
of GGTT that is not assigned to it.

v2: compile fixup
v3: stop using gsm in pte_gpa_not_modifiable_check_via_pf

Signed-off-by: Piotr Piórkowski <piotr.piorkowski@intel.com>
Cc: Michał Wajdeczko <michal.wajdeczko@intel.com>
Cc: Michał Winiarski <michal.winiarski@intel.com>
Cc: Tomasz Lis <tomasz.lis@intel.com>
Reviewed-by: Michał Wajdeczko <michal.wajdeczko@intel.com>