drm/i915/gem: Don't pass uninitialiased execobj to kvfree()
On the error path for relocation handling, the relocs are freed. This
path now includes the internal shadow execobj for cmdparsing that do not
have any relocations defined, and the pointer was left uninitialised.
Use the eb->buffer_count of the relocs that we did copy from before we
added a couple more shadow objects.

Closes: https://gitlab.freedesktop.org/drm/intel/-/issues/2415
Fixes: c43ce12328df ("drm/i915: Use per object locking in execbuf, v12.")
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Tvrtko Ursulin <tvrtko.ursulin@linux.intel.com>