INTEL_DII/NOT_UPSTREAM: drm/i915/ulls: Workaround for GPU hang on BCS0
On XEHPSDV GPU hang on BCS0 is seen while running the ULLS OCL tests
(VLK-20104). The heartbeat request is hung and when heartbeat is
disabled, some other request (engine park or some other request
on BCS0) hangs.

It is observed that GPU hang is not seen if BCS0 engine is kept
active and never parked.OA

So, as a workaround, disable power management for BCS0 on each
tile when ulls contexts are present. Module parameter
ulls_bcs0_pm_wa controls it and is enabled by default.

v2: multi-tile support

Acked-by: Prathap Kumar Valsan <prathap.kumar.valsan@intel.com>
Signed-off-by: Niranjana Vishwanathapura <niranjana.vishwanathapura@intel.com>