INTEL_DII: drm/i915/vf: Clear GT clock frequency during VF removal
During of driver deinitialization, the GT clock frequency is verified
several times.
If this deinitialization is due to a removal of VF, we can expect
DMESG-WARN during this verification, because we are clearing the values of
registers related to GT clock frequency settings.
To avoid this behavior, we can clear the value of gt->clock_frequency, when
we remove the values of registers needed to verify clock frequency.
The next step is to disallow verification when gt->clock_frequency is equal
zero.

Jira: VLK-16596

Signed-off-by: Piotr Piórkowski <piotr.piorkowski@intel.com>
Cc: Michal Wajdeczko <michal.wajdeczko@intel.com>
Cc: Chris P Wilson <chris.p.wilson@intel.com>
Reviewed-by: Chris Wilson <chris@chris-wilson.co.uk>