drm/i915/gt: Limit redzone checking of the context image
Currently, everytime a context is retired we check the tail of the
context image to see if the HW overwrite the following page (as we have
hardcoded the context image size and need to verify that we have
provided sufficient space for HW in its current configuration). On llc,
this check is barely noticeable, but if we need to read the image back
using WC or over PCIe, the cost becomes exhorbitant, even for debugging.

We take two steps here to limit the runtime cost of checking the context
image. Firstly, we only perform the check every 5 seconds, but at least
once for the lifetime of each context. And secondly, instead of checking
the whole page we sample 2 cachelines and use streaming loads to
reduce the cost of reading back over WC.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Cc: Matthew Auld <matthew.auld@intel.com>
Reviewed-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>