drm/i915: Type check i915_request.fence.seqno
We use only a 32b seqno, as that is what is consistently supported across
all our generations of HW, and over all instructions. However, the
dma_fence has expanded its type to a u64, and so when we compile with
-Wconversions, gcc flags all the implicit truncations of the seqno into
the u32 as potential errors. Make those casts implicit, and include an
type overflow check for undefined behaviour.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Brian Welty <brian.welty@intel.com>

(Rebased after mm lock removal patch series)