INTEL_DII/NOT_UPSTREAM: drm/i915: Add support for 1G mapping
When an object gets a 1G chunk of contiguous memory we could take
advantage of it and map the whole 1G chunk in pde, without having to
allocate lower-level entries in pagetable walk.
This patch is reviving an older patch from upstream that was reverted
internally because at that there were no users of 1G allocation.

Revert: 95d896f20131631428e078e4b6ccb8f556e34644
Send 1G-GTT-pages back to the shadow-realm

Signed-off-by: Venkata Sandeep Dhanalakota <venkata.s.dhanalakota@intel.com>
Reviewed-by: CQ Tang <cq.tang@intel.com>