INTEL_DII/NOT_UPSTREAM: drm/i915/hwconf: Add support for hwconfig overrides
There may be requirements for overriding the hwonfig tables to either
add or replace entries. For example, during PO when a rapid turn
around is required and getting a new signed table into a new GuC
firmware will take too long. Or as a temporary workaround until other
software has been updated to match a change in the table that had been
missed.

Signed-off-by: John Harrison <John.C.Harrison@Intel.com>
Reviewed-by: Matthew Brost <matthew.brost@intel.com>