INTEL_DII: drm/i915/gt: Exclude first lmem page for Wa_1409502670
Wa_1409502670 requires all execution on the GPU to avoid using ppGTT
address 0. Since we map the lmem region directly into the kernel's VM
(lmem offset 0 -> VM address 0), in order to satisfy the w/a we need to
exclude allocation of the first lmem page. We could reserve that page for
use with some kernel internal that will never be accessed from the kernel
VM (e.g. the kernel context), but that is rather more complicated.

Signed-off-by: Chris Wilson <chris.p.wilson@intel.com>
Reviewed-by: Matthew Auld <matthew.auld@intel.com>