INTEL_DII: drm/i915: Ensure that the buddy allocator returns aligned offsets
Having the chunk alignment of the buddy allocator translate into natural
physical alignment in the underlying backing device is a very useful
property. Let's preserve this property even if the backing store start is
not naturally aligned with the buddy's chunks by discarding any small
misalignment at the start.

v2: Randomize the selftest and fix the leaks if the st fails.

Suggested-by: Matthew Auld <matthew.auld@intel.com>
Signed-off-by: Chris Wilson <chris.p.wilson@intel.com>
Reviewed-by: Matthew Auld <matthew.auld@intel.com>