INTEL_DII: drm/i915: Pass the memory region start/end into the buddy allocator
Now the buddy allocator allows for the starting offset to be specified,
we can ask the buddy to return the dma_addr_t of each block. This has the
advantage of ensuring that every block returned from the buddy has a
naturally aligned dma_addr_t irrespective of any offset applied to the
dma base. A 1G huge page allocation will therefore be aligned to 1G and
suitable for use as a huge page in the ppGTT.

The caveat is that we have to remember that the memory region's buddy is
absolute and when reserving range we are using relative offsets.

Suggested-by: Matthew Auld <matthew.auld@intel.com>
Signed-off-by: Chris Wilson <chris.p.wilson@intel.com>
Reviewed-by: Matthew Auld <matthew.auld@intel.com>