INTEL_DII: drm/i915: Store a drm_i915_file_private backpointer on the client
The separation between the client and the GEM api grows thinner as we try
to expose more details of GEM through the agnostic client interface. As
such, we will need to walk some details within the GEM file to describe
them to an attached debugger. While we should be extracting the uapi
concepts for debugging, in the meantime let's track which
drm_i915_file_private is associated with the client in order to peek
within.

Signed-off-by: Chris Wilson <chris.p.wilson@intel.com>
Reviewed-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>
Reviewed-by: Andi Shyti <andi.shyti@intel.com>