drm/i915/gt: Maintain rc6-residency counters cross S0 (suspend-to-idle)
During suspend-to-idle, the PCI device is not powered down and so we
retain register state. During this suspend, we only idle the HW and keep
it in rc6 for the maximum amount of powersaving we can achieve with
powergating. This means that we expect rc6 residency to increase during
the suspend, and we must be careful not to confuse ourselves with the
register state being retained.

Currently, we expect the registers to be cleared upon entry to
intel_rc6_sanitize() and clear rc6->prev_hw_residency to match.
Otherwise, on the first sampling, the registers will be less than the
previous reading before suspend and so we will assume that there has
been a wrap around and advance the culmulative total accordingly.
However, across S0, the registers are not reset, but
rc6->prev_hw_residency was, so we would jump the culmulative forward by
the entire residency since boot; that is we would report that during
suspend we spent more time in rc6 than walltime would allow for.

We avoid this overcounting, while still resetting the counters to match
the registers across S3/S4, by moving the memset() from sanitize to
intel_rc6_disable().

Testcase: igt/i915_pm_rc6_residency/rc6-freeze
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Tvrtko Ursulin <tvrtko.ursulin@linux.intel.com>
Cc: Mika Kuoppala <mika.kuoppala@linux.intel.com>
Reviewed-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>