INTEL_DII: drm/i915: Add return code for put_pages() and misc swapping fixes.
We use i915_window_blt_copy() to do the blt swapping copy, this call might get
interrupted, so we need to return -ERESTARTSYS or -EINTR back to userspace.
These swapping operations are under put_pages() call. Therefore put_pages() is
changed to return error code.

i915_gem_object_memcpy() does not check if the objects are with different size,
we need to use the smaller size.

We fallback to memcpy based swapping if there is no flat CCS on the platform.