drm/i915/gt: Remove stale breadcrumb assertion
Since reordering the i915_request_retire vs
i915_request_enable_breacrumb synchronisation points, it is no longer
true that the breadcrumb is fully inserted before the final signaling in
i915_request_retire.

During i915_request_enable_breadcrumb on the submission, the reference
to the request is owned by the submission path and released during
i915_request_retire. Therefore we need to make sure that we have
serialised acquisition of the reference for the use by the breadcrumb
signaling prior to releasing the reference during retire. The
synchonisation lock used for this is the engine->active.lock, being held
by the submission to insert the request onto the engine's queue, and
then taken by retire to remove the signaled request from that queue.
Ergo retire has to wait for the completion of the request's submission
before it can release the reference, and so we can safely acquire a new
reference for the breadcrumb.

Similarly, it is safe for us to insert a signaled request onto the irq
worker's queue so long as we hold references to that request and
context; it is simply extra work to then remove the already signaled
request during the irq work. But that is no different than if the
request were to be signaled by any other means.

v2: Clarify why we use rq->lock to serialise between
dma_fence_enable_signaling / i915_request_enable_breadcrumbs() and
i915_request_retire()

Closes: https://gitlab.freedesktop.org/drm/intel/-/issues/3687
Signed-off-by: Chris Wilson <chris.p.wilson@intel.com>
Cc: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Reviewed-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>