drm/i915: Soften module failure on memory region setup failure
If the memory region setup fails, log an error, and continue on. If we
have no essential memory regions, fail outright, otherwise try and limp
on with reduced functionality. The goal is that we do complete the
module load and be able to report the failure cleanly, and facilitate
the user updating their system (i.e. we aim to support minimal kms).

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Matthew Auld <matthew.auld@intel.com>
Reviewed-by: Matthew Auld <matthew.auld@intel.com>