INTEL_DII/UAPI/NEEDSIGT: drm/i915/debugger: Add ack events for vm_bind
Before we can deliver uuids associated with vm_binds into
the hardware, debugger needs to know their location (for
modification like breakpoints).

In direct submission for a debugger have any means to affect
the ISA contents before workload is getting executed, we need to
halt the client on vm_bind and let the debugger intercept
at this moment (on immediate binds).

Add a fence to the vma associated with vmbind and
and wait for it in submission or in the ioctl.

Signal fence by requiring vm bind create events
with uuids to be acked by the debugger.

For IMMEDIATE binds we will block in the client
vm_bind_ioctl and for non immediate ones
we block in the submission.

Vma fence parts are based on snippet from Chris Wilson.

Jira: VLK-9252

Signed-off-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>