drm/i915/gt: Force a context switch around vm reconfiguration
On gen8-gen11, updates to the page-directory registers are immediate,
and future address translations use the new page tables. On gen12, the
page directory is not reloaded until after a context switch (or at least
a PD reload through the ELSP). With execlists, requesting a context
switch around a request is possible with the SENTINEL bit, but to work
with the guc submission, we need to manually perform a context switch
from which to reprogram the registers.

As this was the only user of the context-barrier, we can replace it
entirely.

v2: Downstream return void from context_apply_all (Pallavi)

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: CQ Tang <cq.tang@intel.com>
Reviewed-by: Andi Shyti <andi.shyti@intel.com>