INTEL_DII/FIXME: drm/i915/gt: Fake interrupts for Wa:16014207253
All current multi-tile hardware appears to have a bug where a tile 1
to tile 0 communication channel gets confused with posted writes and
locks up. That results in a total hard hang of the GPU. One way to
avoid this is to not send posted writes across this channel. The major
user of which is interrupts - both G2H and USER_INTERRUPT.

Hence the workaround is to not enable interrupts at all. This
patch changes the driver to use a polling mechanism instead. The
polling frequency is 10Hz when nothing is expected (but asynchronous
G2H can still be sent - engine hang notifications and such). When the
GPU is busy (outstanding requests or G2H responses), the frequency is
pushed to 10kHz. Initial testing by compute and media teams suggests
this frequency is the best compromise.

The w/a is enabled for all affected platforms. In theory, it could be
limited to the 2nd tile only. The decision from sw arch was to
implement it for all as not worth the extra complication to do single
tile only.

FIXME: Currently the polling is enabled the whole time that interrupts
would be enabled. It could be tied to the GT PM enable instead.
However, that requires fixes to the GT PM management with respect to
G2H traffic which are in development.

FIXME: The fake interrupt handlers are not currently implemented for
execlist mode. So if a multi-tile platform is run in execlist mode, it
won't work - no interrupts and no polling.

v2: Move into XeHPSDV section of tree and re-instate debugfs controls
after they were removed by an upstream rebase (JohnH).
v3: Rebase from GuC engine busyness backport (Umesh)
v4: Disable polling while sending CLIENT_RESET to prevent spurious
GUC_CTB_STATUS_UNUSED failures (JohnH)

Signed-off-by: John Harrison <John.C.Harrison@Intel.com>