drm/i915/gem: Translate E2BIG to VM_FAULT_SIGBUS
As we may defer checking whether an object can fit within the
constraints of the backing store to allocation from that store, we may
see E2BIG anytime we first retrieve the pages for an object. It is an
expected error we need to handle during the faulthandler, with the
appropriate response of sending SIGBUS to the client as the object is
invalid and may never be mmapped.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Matthew Auld <matthew.auld@intel.com>
Reviewed-by: Matthew Auld <matthew.auld@intel.com>