INTEL_DII/NOT_UPSTREAM: Enable scratch for fault handler
When we create a vm, we map its entire address space to a scratch
page in ppgtt. Any user workload accessing an invalid address would
result in accessing the scratch page. Without the scratch page setup any
such access would have result in catastrohic fault requiring  i915 to
kill the context.

With graphics recoverable page fault supported from pvc, we cannot
enable this scratch page setup when creating a vm because we bind the
vma in the fault handler and if we had mapped the entire address space
to the scratch we would never receive a page fault.

However UMD is relying on having scratch. We are asked to
provide a mechanism when faults are enabled to map any invalid
access to scratch and report it to UMD instead of triggering a
catastrophic fault. This will let UMD to continue and also provides
invalid address information to aid in their debugging to fix all invalid
accesses from UMD.

In this patch, if fault handler receives a fault for an invalid access,
it maps that address to the scratch page. We also need to kill this
scratch mapping when a new vma is created, because followed by an
invalid access, that address range could become valid.
We also need to invalidate the tlbs because now the scratch mapping is
invalid.

Reference VLK-21784

v2: Add debug log for invalid va
v3: When increase PD/PT, need to reset next level PD/PT entry (Bruce)
v4: Use Null page for scratch page (Bruce)
v5: Invalidate TLB only if scratch page is used to map invalid VA.
    Currently TLB invalidation is performed on vm_bind unconditionally
    if page fault is enabled. This has caused 26% performance drop
    for some benchmark, compared to page fault disabled. See more
    details in VLK-44706. Fix this issue by invalidating scratch
    pages conditionally.
    While here, also remove TLB invalidation for the cases where VA
    range is mapped to invalid scratch pte. HW doesn't cache invalid
    PTE so no need to invalidate TLB for those cases. (Oak)

Signed-off-by: Prathap Kumar Valsan <prathap.kumar.valsan@intel.com>
Reviewed-by: CQ Tang <cq.tang@intel.com>