INTEL_DII/FIXME: drm/i915/mtl: add GSC CS interrupt support
The GSC CS re-uses the same interrupt bits that the GSC used in older
platforms. This means that we can now have an engine interrupt coming
out of OTHER_CLASS, so we need to handle that appropriately.

FIXME: the code currently considers the old GSC interface and the
new GSC CS as separate things and checks for both. The code should be
refactored to limit the duplication.

v2: mark pick_gt() as static, drop duplicated comment (Matt)

Bspec: 65371
Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Cc: Matt Roper <matthew.d.roper@intel.com>
Reviewed-by: Matt Roper <matthew.d.roper@intel.com>