INTEL_DII/UAPI/NEEDSIGT: drm/i915/debugger: Mark intel_context with a debugger id
Intel engines are now equipped with a new debugger id that
identifies the Logical Ring Context (LRC) during scheduling.

The lrc is then reported to the debugger that will keep track of
the tuple "context id"/"lrc id". This is done by generating a new
event PRELIM_DRM_I915_DEBUG_EVENT_ENGINES that reports the
list of lrc handles (including class/instance) associated to the
context. The event is sent after a
PRELIM_DRM_I915_DEBUG_EVENT_CONTEXT_PARAM if there are intel
engines created along with the context.

  Note: a natural place where to add the list of LRCs would have
  been in the context_param structure, but a new event has been
  added instead. This because adding an array of intel_engines
  requires a flexible array in the context_param structure which
  is already already flexible (see ctx_param->param.value).
  Adding a second flexible array required some tricks that would
  affect the API specification.

When an eu attention is sent, the event contains the lrc handle.
The debugger at this point knows how to get to the context.

For now both the context id and lrc id are saved in the eu
attention event so that the debugger team has time to transition
to the new architecture.

Signed-off-by: Andi Shyti <andi.shyti@intel.com>
Cc: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
Cc: Mika Kuoppala <mika.kuoppala@linux.intel.com>
Reviewed-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>