INTEL_DII: drm/i915/debugger: Add eu and l3 invalidates
If debugger does a pread from a vm, we need to flush l3
state before actual read to get the memory contents the
engine did use.

If debugger does a pwrite from a vm, we need to invalidate
the eu's so that on continuing from a breakpoint the
modifications made by debugger are actually seen by
the engine.

v2: Forcewake gt and render domains during the invalidation
v3: Mutex instead of spinlock during eu flush

Jira: VLK-11248

Cc: Andi Shyti <andi.shyti@intel.com>
Signed-off-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>
Signed-off-by: Tilak Tangudu <tilak.tangudu@intel.com>
Acked-by: Maciej Patelczyk <maciej.patelczyk@intel.com>
Acked-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>