drm/i915: Avoid reloading signal->engine during __await_execution()
During submission of a new request, we may be looking at the execution
status of a signaler (to evaluate whether we need additional scheduling
points between the two). As the execution status of the signaler is in
flux (it is being run independently by the HW), we have to be careful
that the state may change during the read. In particular, with a signal
along a virtual engine, between checking that the signal is executing
(intel_context_inflight) and inspecting which engine it is running on,
it may be transferred from one to another. We can avoid reloading the
transient signal->engine, by using the inflight engine returned by
intel_context_inflight.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>